% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cafeF.R
\name{get_cafeF}
\alias{get_cafeF}
\title{Get stock data from cafef.vn}
\usage{
get_cafeF(symbol, start_date = NULL, end_date = NULL)
}
\arguments{
\item{symbol}{the code of stock, for example 'VCB', 'ACB'}

\item{start_date}{the start date, in the form DD/MM/YYYY}

\item{end_date}{the end date, in the form DD/MM/YYYY}
}
\value{
A tibble contain all stock data in that period
}
\description{
get_cafeF() is normally slower then get_vndirect(), because it trying to scrape data instead of
using the behind API
}
