% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cafeF2.R
\name{get_cafeF2}
\alias{get_cafeF2}
\title{Get stock data from cafef.vn with asynch request}
\usage{
get_cafeF2(symbol, start_date = NULL, end_date = NULL)
}
\arguments{
\item{symbol}{the code of stock, for example 'VCB', 'ACB'}

\item{start_date}{the start date, in the form DD/MM/YYYY}

\item{end_date}{the end date, in the form DD/MM/YYYY}
}
\value{
A tibble contain all stock data in that period
}
\description{
get_cafeF2() is now using asynchronous request from `crul` package to improve speed. The function now is
twice as fast as the old one.
}
